dtoc: Add the method for each command to OutputFile
Rather than the if/else construct, update OutputFile with the method to
call to process each command. This is easier to maintain as the number of
commands increases.

Rename generate_tables to generate_plat since it better describes what is
being generate ('plat' is the U-Boot name for platform data).

Signed-off-by: Simon Glass <sjg@chromium.org>