dtoc: Support scanning of structs in header files
Drivers can private / platform data contained in structs and these struct
definitions are generally kept in header files. In order to generate
build-time devices, dtoc needs to generate code that declares the data
contained in those structs. This generated code must include the relevant
header file, to avoid a build error.

We need a way for dtoc to scan header files for struct definitions. Then,
when it wants to generate code that uses a struct, it can make sure it
includes the correct header file first.

Add a parser for struct information, similar to drivers. Keep a dict of
the structs that were found.

Signed-off-by: Simon Glass <sjg@chromium.org>