dtoc: Support tracking the phase of U-Boot
U-Boot operates in several phases, typically TPL, SPL and U-Boot proper.
The latter does not use dtoc.

In some rare cases different drivers are used for two phases. For example,
in TPL it may not be necessary to use the full PCI subsystem, so a simple
driver can be used instead.

To make this work we need to be able to mark drivers with their phase.
This is done by adding a macro to the driver. Add support for this in
dtoc.

Signed-off-by: Simon Glass <sjg@chromium.org>