dm: core: Allow dropping run-time binding of devices
With OF_PLATDATA_INST devices are bound at build time. We should not need
binding of devices at runtime in most cases. However it is inflexible to
absolutely prohibit it, so add an option to control this.

Update the driver model core so that it does not bind devices. Update
device_bind() to return an error if called.

Series-changes: 3
- Update the commit message to explain OF_PLATDATA_NO_BIND better
- Update OF_PLATDATA_NO_BIND help as well

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I2883cc0cb1edf613bf2dfb30e4bf982cf03daafc