smbios: Drop the eos parameter
We can store this in the context and avoid passing it to each function.
This makes it easier to follow and will also allow keeping track of the
end of the string table (in future patches).

Add an 'eos' field to the context and create a function to set it up.

Series-changes: 3
- Fix comment for smbios_add_prop()
- Rename set_eos() to smbios_set_eos()

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I6c28d884e356bafbb28c4fa3a1ce66b9fa761158
Reviewed-by: Bin Meng <bmeng.cn@gmail.com>