smbios: Allow a few values to come from sysinfo
While static configuration is useful it cannot cover every case. Sometimes
board revisions are encoded in resistor straps and must be read at
runtime.

The easiest way to provide this information is via sysinfo, since the
board can then provide a driver to read whatever is needed.

Add some standard sysinfo options for this, and use them to obtain the
required information.

Series-to: u-boot
Series-cc: x86-colab
Cover-letter:
smbios: Enhancements for more flexibility
This series includes various patches to allow more flexibility as to where
the data for SMBIOS tables comes from:

- introduces some standard sysinfo options as a source, e.g. to read
   strapping pins to determine the board revision
- allows the U-Boot version number to be included
- allows the version number to be provided programmatically, e.g. to
   support the build system adding information after U-Boot is built

Documentation is added for how to obtain version information.

The code is also refactored a little to make it easier to maintain.
END

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I2ef14d4a29ee813c1a6c3577d10df8fe6062172d