binman: Allow for skip_at_start when reading entries
The offset of an entry needs to be adjusted by its skip-at-start value.
This is currently missing when reading entry data. Fix it.

Series-to: u-boot
Cover-letter:
binman: Various minor fixes and improvement
This series includes a number of minor fixes and improvements identified
while trying out Chrome OS verified boot on it.

These were not apparent without this real-world testing, but all are
deficiencies in one way or another.
END

Signed-off-by: Simon Glass <sjg@chromium.org>