log: Convert log values to printf() if not enabled
At present if logging not enabled, log_info() becomes a nop. But we want
log output at the 'info' level to be akin to printf(). Update the macro to
pass the output straight to printf() in this case.

Series-to: u-boot
Series-cc: heinrich
Cover-letter:
log: Allow multiple lines and conversion to printf()
At present when logging is not enabled, all log calls become nops. This
does not seem right, since if the log level is high enough then there
should be some sort of message. So in that case, this series updates it to
print the message if the log level is above LOGL_INFO.

Also the current implementation does not support multiple log calls on the
same line nicely. The tags are repeated so the line is very hard to read.
This series adds that as a new feature.
END
Signed-off-by: Simon Glass <sjg@chromium.org>