sandbox: Rename GPIO dir_flags to flags
Adjust the terminology in this driver to reflect that fact that all flags
are handled, not just direction flags.

Create a new access function to get the full GPIO state, not just the
flags. Drop the static invalid_dir_flags since we can rely on a segfault
if something is wrong.

Signed-off-by: Simon Glass <sjg@chromium.org>