tpm: Allow disabling platform hierarchy with TPM2
With TPM2 we don't actually lock the TPM once verified boot is finished.
Instead we disable the platform hierarchy which serves the same purpose.
Add an implementation of this so we can safely boot into the kernel.

Series-changes: 2
- Add definition of TPM2_RC_NV_DEFINED return code

Series-to: u-boot
Series-links: 226046 2:224784
Series-version: 4
Series-cc: Ilias Apalodimas <ilias.apalodimas@linaro.org>
Cover-letter:
tpm: Support using TPM1 and TPM2 from a single API
At present if an application wants to be written so it can work with
both TPMv1.2 and TPM2 it must use two different APIs. This is inconvenient
since it requires adding code to deal with the mismatch between the two.
It would be better to have a common API that all boards could share.

This series provides a simple API the covers some basic features and
implements them for both TPM standards.
END
Signed-off-by: Simon Glass <sjg@chromium.org>

Change-Id: I151cb828b53c98409ab85c3ea82c774fc6cb88f7
Acked-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>