cros_ec: Add vstore support
The EC can store small amounts of data for the benefit of the
verified boot process. Since the EC is seldom reset, this can allow the
AP to store data that survives a reboot or a suspend/resume cycle.

Add support for this.

Series-to: u-boot
Cover-letter:
cros_ec: Add features required for verified boot
The Chromium OS EC has developed a few more features. This series adds
support for some of these, to allow a more recent verified-boot
implementation to work.

Also some existing features don't have commands available. It is useful
to be able to check the EC state from the command line, so this series
adds commands for listing features, checking switches, etc.

The following parts of the implementation are added or expanded:
   - SKU ID for finding out basic factory information
   - 'features' for seeing what the EC supports
   - 'switches' for LID open, etc.
   - events for seeing when the EC has an event for the AP
   - vstore (for storing data in the EC)

The current comment style is used for new functions.
END

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I3acf20d585d8985e7f8d8aa579429b21d7d452aa