doc: Add a note about producing 'md.b' output using hexdump
Comparing a hex dump on the U-Boot command line with the contents of a
file on the host system is fairly easy and convenient to do manually if
it is small. But the format used hexdump by default differs from that
shown by U-Boot. Add a note about how to make them the same.

(For large dumps, writing the data to the network with tftpput, or to a
USB stick with ext4save is easiest.)

Series-changes: 4
- Indicate that hd does not show the offset correctly
- Add an xxd + sed example too

Series-changes: 3
- Add this to the docs for the md command

Signed-off-by: Simon Glass <sjg@chromium.org>