s5p4418_nanopi2: Bring in some dead code
This code is still using the old command typedef. It was not noticed since
this file is not currently built. It is using a non-existent option in the
Makefile.

Fix it up and add it to the build.

(This is offered as an act of service in the hope of receiving a free
beer at some point.)

Patch-cc: Stefan Bosch <stefan_b@posteo.net>
Signed-off-by: Simon Glass <sjg@chromium.org>