test: Use a local variable for test state
At present we use a global test state for all driver-model tests. Make use
of a local struct like we do with the other tests.

To make this work, add functions to get and set this state. When a test
starts, the state is set (so it can be used in the test). When a test
finishes, the state is unset, so it cannot be used by mistake.

Series-changes: 3
- Use test_set_state() throughout test-main.c instead of direct assignment

Signed-off-by: Simon Glass <sjg@chromium.org>