RFC: display_options: Use print_hex_dump() for print_buffer()
These two functions do similar things. When CONFIG_HEXDUMP is enabled,
drop the code in print_buffer() and use the hexdump code instead. This
increases the code size a little, but makes the API similar to Linux.

When CONFIG_HEXDUMP is not enabled, don't do this, since presumably
code size is more important.

To make this work, update the address test so that the address matches
the pointer, since the hexdump routine does not support an arbitrary
address.

This is not a great result, but it is a step towards unifying the APIs.
I doublt we want this patch, which is why it is marked RFC. It might be
better to unify the other way, i.e. reimplement the hexdump routines.