tpm: Allow commiting non-volatile data
Add an option to tell the TPM to commit non-volatile data immediately it
is changed, rather than saving it for later.

Add definitions for the rest of the Cr50 commands while we are here.

Series-to: u-boot
Series-cc: tpm
Cover-letter:
tpm: Various minor fixes and enhancements
This series contains some minor enhancements for the TPM code to make it
work with Chromium OS verified boot.
END

Signed-off-by: Simon Glass <sjg@chromium.org>