pxe: Clean up the use of bootfile
The 'bootfile' environment variable is read in the bowels of pxe_util to
provide a directory to which all loaded files are relative.

This is not obvious from the API to PXE and it is strange to make the
caller set an environment variable rather than pass this as a parameter.

Update the API to improve this. Unfortunately this means that
pxe_setup_ctx() can fail, so add error checking.

Change-Id: I6a01e73ef114448e39cb6899c21f6c169e4da216
Signed-off-by: Simon Glass <sjg@chromium.org>