bootstd: Add the bootdev uclass
A 'bootdev' is a device which can be used to boot an operating system.
It is a child of the media device (e.g. MMC) which handles reading files
from that device, such as a bootflow file.

Add a uclass for bootdev and the various helpers needed to make it
work. Also add a binding file, empty for now.

Series-changes: 3
- Update for the new trailing_strtoln_end()

Change-Id: I3b5bc09d034cd72b2a9c604a0907c10abc05956a
Signed-off-by: Simon Glass <sjg@chromium.org>