bootstd: Add tests for bootstd including all uclasses
Add a set of combined tests for the bootdev, bootflow and bootmeth
commands, along with associated functionality.

Expand the sandbox console-recording limit so that these can work.

These tests rely on a filesystem script which is not yet added to the
Python tests. It is included here as a shell script.

Series-changes: 3
- Update bootdev_test_order() for boot_targets env var
- Add tests for the "bootmeths" env var
- Update tests for the new 'bootmeth order' syntax

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I936bf6fe715bd1feac569478aded46ab75579148