x86: Create a 32/64-bit selection for the app
Most EFI implementations use 64-bit. In order to spupose a 64-bit app,
update the Kconfig to add an option for 32/64 bit. Update the prompt for
the existing option so it is clear it relates to the stub. Move both
up to just under the choice that controls them, since this looks better
and the menu.

Use CONFIG_EFI_APP in the Makefile instead of CONFIG_TARGET_EFI_APP,
since the latter is specific to a single target and we will have two.

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Id32ab4eff4b2a6b82fed22813a01fc2a08c6e0a5