efi: Add video support to the app
The current EFI video driver only works when running in the stub. In that
case the stub calls boot services and copies the graphics info over to the
efi table. This is necessary because the stub exits boot services before
jumping to U-Boot.

The app maintains access to boot services throughout its life, so does not
need to do this. Update the driver to support calling boot services
directly.

Enable video output for the app.

A sample qemu command-line for this case is:

   qemu-system-x86_64 -bios /usr/share/edk2.git/ovmf-ia32/OVMF-pure-efi.fd
   -drive id=disk,file=try.img,if=none,format=raw -nic none
   -device ahci,id=ahci -device ide-hd,drive=disk,bus=ahci.0

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ibf877904b58b5ba17bdd00f80dcd3f8b8e0c1f39