efi: Allow easy selection of serial-only operation
Add info about how to select vidconsole or serial.

Series-to: u-boot
Series-cc: trini, heinrich, ilias,
Series-cc: Christian Melki <christian.melki@t2data.com>
Cover-letter:
efi: Improvements to U-Boot running on top of UEFI
At present U-Boot can be built as an EFI app, but it is really just for
testing, with very limited features. Typically the payload build is used
to boot on top of UEFI, where U-Boot takes over the machine immediately
and supplies its own drivers.

But the app could be made more useful.

This series provides access to EFI block devices and the video console
within U-Boot. It also restructures the app to make it possible to boot
a kernel, although further work is needed to implement this.

This can be thought of as making U-Boot perform a little like 'grub', in
that it runs purely based on UEFI-provided services.

Of course a lot more work is required, but this is a start.
END

Change-Id: I8927cdce4b36339b962ef31074500198fcc3be1c
Signed-off-by: Simon Glass <sjg@chromium.org>