efi: Add EFI uclass for media
At present UCLASS_EFI is used to represent an EFI filesystem The
description of this uclass is "EFI managed devices" which is pretty vague.
The only driver that uses this uclass is in fact not a real U-Boot driver,
since its operations do not include a struct udevice.

Rather than mess with this, create a new UCLASS_EFI_MEDIA uclass to handle
EFI media such as disks. Make this the uclass to use for EFI media.

The existing implementation using UCLASS_EFI remains as is, for
discussion.

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I50ca59e389adf282bbff50eecc593f010a89c0f7