WIP: bloblist: Support allocating the bloblist
Typically the bloblist is positioned at a fixed address in memory until
relocation. This is convenient when it is set up in SPL or before
relocation.

But for EFI we want to set it up only when U-Boot proper is running.

WIP: Needs test and docs
WIP: Should reloc_bloblist() do anything in this case?

Signed-off-by: Simon Glass <sjg@chromium.org>