Makefile: Include the config for the phase being built
At present there is only a single auto.conf file used within the
makefiles. Update them to use the correct one for each phase.

Also update kconfig.h to include the correct autoconf.h or autoconf_spl.h
file for each phase. This allows the macros to be simplified.

With this, CONFIG_IS_ENABLED() is the same as IS_ENABLED() apart from a
migration detail.

Both changes are made in the same patch to avoid build errors.

Series-changes: 2
- Include code from 'Update CONFIG_IS_ENABLED() for split files' too

Signed-off-by: Simon Glass <sjg@chromium.org>