bootstd: Add setup for the bootflow tests
We need to create a disk image with a partition table and a DOS-format
filesystem containing a few files. Provide a fallback binary for CI since
it does not seem able to detect the loopback partitions.

Add this to a dm_init test so that it happens when needed.

Series-changes: 3
- Update test condition to use CMD_BOOTFLOW

Signed-off-by: Simon Glass <sjg@chromium.org>