binman: Write fake blobs to the output directory
At present binman writes fake blobs to the output directory. This is not
very helpful, since the files serve no useful purpose. Write them to the
output directory instead.

Move the file-creation code to the Entry base class, so it can be used by
any entry type that needs it. This is required since some entry types,
such as Entry_blob_ext_list, are not subclasses of Entry_blob.

Signed-off-by: Simon Glass <sjg@chromium.org>