patman: Allow running a tool and returning the full result
Add a new function which returns the entire result from running a tool,
not just stdout. Update Run() to use thism and to return stdout on error,
if stderr is empty, since some tools write error output to stdout.

Make the exception catching more specific, to just ValueError, since
broad exceptions are a pain to debug.

Signed-off-by: Simon Glass <sjg@chromium.org>