arm: qemu: Add an SPL build
Add an SPL build for qemu so we can test the standard passage feature.

Include a binman definition so that SPL and U-Boot are in the same image.
This requires adding a proper devicetree file for qemu_arm. It is only
used for the SPL build.

For now this just boots to a prompt.

Signed-off-by: Simon Glass <sjg@chromium.org>