tpm: sandbox: Allow init of TPM in a different phase
At present the emulator assumes that the TPM is inited in the same phase
where it is used. But in fact SPL may init the TPM, so we don't want to
complain when U-Boot proper later uses it. Remove this check.

It might be best to save this information into the device state for the
TPM, so that we can make sure the TPM was inited at some point. For now,
this seems good enough.

Series-to: u-boot
Cover-letter:
sandbox: Various minor fixes and enhancements
This series contains some minor enhancements for sandbox to make it work
with Chromium OS verified boot.
END

Signed-off-by: Simon Glass <sjg@chromium.org>