vbe: Add a new vbe command
Add a command to look at VBE methods and their status. Provide a test for
all of this as well.

Series-to: u-boot
Series-cc: trini
Series-links: 310481
Cover-letter:
vbe: Verified Boot for Embedded initial support
This adds the concept of a VBE method to U-Boot, along with an
implementation of the 'VBE simple' method, basically a simple way of
updating firmware in MMC from userspace and monitoring it from U-Boot.

VBE simple is implemented in fwupd. U-Boot's role is to set up the
device tree with the required firmware-update properties and provide the
developer with information about the current VBE state. To that end this
series includes a new 'vbe' command that allows VBE methods to be listed
and examined.

As part of this work, support for doing FDT fixups via the event interface
is provided, along with the ability to write to the device tree via the
ofnode interface.

Further work is needed to pull everything together, but this is a step
along the way.
END

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I0ec5199bc0e81fe308317b055c69c48a5135d2f7