Revert "fdtdec: drop needlessly convoluted CONFIG_PHANDLE_CHECK_SEQ"
The fdt_path_offset() function is slow since it must scan the tree.
This substantial overhead now applies to all boards.

The original code may not be ideal but it is fit for purpose and is only
needed on a few boards.

Reverting this reduces time to set up driver model by about 30ms.

Before revert:

Accumulated time:
                47,170  dm_r
                53,237  dm_spl
               572,986  dm_f

Accumulated time:
                44,598  dm_r
                50,347  dm_spl
               549,133  dm_f

This reverts commit 26f981f295d00351b6f0c69b5317b254b2361cc0.

Series-changes: 1
- Add detail of impact

Series-to: u-boot
Cover-letter:
fdt: Improvements for fdt and tracing
This series aims to make it easier to measure boot performance,
particularly aimed at rockpro64. It includes the following:

- enabling bootstage on rockpro64 (TPL, SPL, U-Boot proper)
- minor fixes and tidy-ups for tracing
- minor FDT code tidy-ups

It also reverts a patch which slows down the boot by 30ms. This revert
was requested some time ago but with this series we have the data showing
the impact.
END

Signed-off-by: Simon Glass <sjg@chromium.org>